refactor: app폴더 아래는 라우팅관련만 위치
## 작업 내용
- actions, dummy 폴더는 app폴더 바깥으로 위치 변경
- nav 메뉴로 url 이동 시 -> category 페이지 작성
